Rename PaymentITData to PaymentIntegrationTestData
