+++ cab-app-payment-service/src/integration-test/java/by/arvisit/cabapp/paymentservice/controller/PassengerPaymentControllerIT.java
@@ -1,18 +1,18 @@
 package by.arvisit.cabapp.paymentservice.controller;
 
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.PASSENGER_PAYMENT_1_ID;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.PAYMENT_METHOD_REQUEST_PARAM;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.URL_PASSENGER_PAYMENTS;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.URL_PASSENGER_PAYMENTS_ID_TEMPLATE;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getAddedPassengerPaymentResponseDto;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getListContainerForResponse;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getPassengerPaymentRequestDto;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getPassengerPaymentResponseDto1;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getPassengerPaymentResponseDto2;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getPassengerPaymentResponseDto3;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getPassengerPaymentResponseDto4;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getPassengerResponseDto;
-import static by.arvisit.cabapp.paymentservice.util.PaymentITData.getRideResponseDto;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.PASSENGER_PAYMENT_1_ID;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.PAYMENT_METHOD_REQUEST_PARAM;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.URL_PASSENGER_PAYMENTS;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.URL_PASSENGER_PAYMENTS_ID_TEMPLATE;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getAddedPassengerPaymentResponseDto;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getListContainerForResponse;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getPassengerPaymentRequestDto;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getPassengerPaymentResponseDto1;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getPassengerPaymentResponseDto2;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getPassengerPaymentResponseDto3;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getPassengerPaymentResponseDto4;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getPassengerResponseDto;
+import static by.arvisit.cabapp.paymentservice.util.PaymentIntegrationTestData.getRideResponseDto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
